requirements.txt
line-bot-sdk
openai
google-cloud-storage
google-cloud-firestore